requirements.txt
selenium
Pillow
opencv-python
pytesseract